requirements.txt
flask==2.2.5
flask_restx
flask_sqlalchemy
flask_jwt_extended
python-dotenv